lpr-api/requirements.txt
fastapi
uvicorn
torch
ultralytics
numpy
opencv-python
Pillow